chore: revert backend changes
