+++ catalog/cmd/catalog/config.go
@@ -14,7 +14,6 @@ type config struct {
 	ShutdownTimeout         time.Duration
 	PluginAddresses         map[string]string
 	PluginConnectionTimeout time.Duration
-	PluginTimeout           time.Duration
 }
 
 type envConfig struct {
@@ -23,7 +22,6 @@ type envConfig struct {
 	ShutdownTimeout         time.Duration `env:"SHUTDOWN_TIMEOUT" default:"5s"`
 	PluginAddresses         []string      `env:"PLUGIN_ADDRESSES"`
 	PluginConnectionTimeout time.Duration `env:"PLUGIN_CONNECTION_TIMEOUT" default:"10s"`
-	PluginTimeout           time.Duration `env:"PLUGIN_TIMEOUT" default:"5m"`
 }
 
 func loadConfig() (config, error) {
@@ -46,7 +44,6 @@ func loadConfig() (config, error) {
 		ShutdownTimeout:         raw.ShutdownTimeout,
 		PluginAddresses:         pluginAddresses,
 		PluginConnectionTimeout: raw.PluginConnectionTimeout,
-		PluginTimeout:           raw.PluginTimeout,
 	}
 
 	return cfg, nil
