+++ catalog/internal/catalog/memory_store.go
@@ -12,7 +12,6 @@ import (
 
 var (
 	ErrInvalidIngestion = errors.New("invalid ingestion payload")
-	ErrNodeNotFound     = errors.New("node not found")
 )
 
 type Store interface {
